requirements.txt
asgiref==3.5.2
cssselect2==0.6.0
Django==4.0.6
djangorestframework==3.13.1
easy-thumbnails==2.8.1
lxml==4.9.1
Pillow==9.2.0
pytz==2022.1
reportlab==3.6.11
sorl-thumbnail==12.8.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
webencodings==0.5.1
psycopg2-binary~=2.8.6
